CampusTrace-Backend/requirements.txt
fastapi
uvicorn[standard]
supabase
python-multipart
pydantic-settings
email-validator
google-generativeai
Pillow
click
h11
PyJWT
protobuf
python-dotenv
httpx
resend
numpy
transformers
torch
backoff
jina
